\name{insurance}
\docType{data}
\alias{insurance}
\title{Insurance cost data for nonsmokers}
\description{
  Total health insurance costs for the nonsmokers in a simulated study.  
}
\usage{insurance}
\format{
  The data consists of the response variable which is total healthcare cost 
  paid by an insurer measured in thousands of dollars (charges) and two 
  predictors which are age in years (age) and body mass index (bmi).  The 
  predictor variables are rescaled so that the support of the predictor space 
  is [0,1]^2.
}
\references{
  Lantz, Brett (2013) 
  \emph{Machine learning with R}, Packt Publishing Ltd. 
  \url{https://www.kaggle.com/lbronchal/explanatory-models-for-healthcare-costs}
}
\keyword{datasets}
